Add one binary unit-test
