+++ tests/suites.h
@@ -1,0 +1,5 @@
+
+#pragma once
+
+int add_ref_count_suite(void);
+
